refactor: 컴파일러 경고 대응
