+++ src/main/kotlin/upbrella/be/rent/dto/response/RentalHistoryResponse.kt
@@ -2,7 +2,6 @@ package upbrella.be.rent.dto.response
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import java.time.LocalDateTime
-import upbrella.be.rent.entity.History
 
 data class RentalHistoryResponse(
     val id: Long,
