+++ src/test/kotlin/upbrella/be/rent/controller/RentControllerTest.kt
@@ -33,7 +33,7 @@ import upbrella.be.rent.service.ConditionReportService
 import upbrella.be.rent.service.ImprovementReportService
 import upbrella.be.rent.service.LockerService
 import upbrella.be.rent.service.RentService
-import upbrella.be.slack.service.SlackAlarmService
+import upbrella.be.slack.SlackAlarmService
 import upbrella.be.user.dto.response.SessionUser
 import upbrella.be.user.entity.User
 import upbrella.be.user.repository.UserReader
