+++ src/test/kotlin/upbrella/be/rent/service/RentServiceTest.kt
@@ -28,7 +28,7 @@ import upbrella.be.rent.exception.NonExistingHistoryException
 import upbrella.be.rent.exception.NotAvailableUmbrellaException
 import upbrella.be.rent.exception.NotRefundedException
 import upbrella.be.rent.repository.RentRepository
-import upbrella.be.slack.service.SlackAlarmService
+import upbrella.be.slack.SlackAlarmService
 import upbrella.be.store.entity.StoreMeta
 import upbrella.be.store.repository.StoreMetaReader
 import upbrella.be.umbrella.entity.Umbrella
